Use the backbuffer size for the RenderTarget dimensions
If I understand correctly, the RenderTarget is used to temporarily proxy
the data that's supposed to be written to the GPU (e.g. to allow us to
inspect it, for example for capturing screenshots or apply filters). Then
the GPU is set to direct mode (write to back buffer with no proxy), and
the data in the RenderTarget gets copied back.

So it makes no sense to have different dimensions for the backbuffer and
the RenderTarget. If the window is full screen or resized according to
player preferences, both the back buffer and render target should change
dimensions to identical values.