fix: physics engine crashing due to NaN when finding player bounding box
This happened when you pressed two opposing movement keys simultaneously,
like left and right or up and down.

This was due to player_vel being Vector3.Zero, yet move_dirty being true,
which caused a zero vector to be normalised (to NaN) and passed to the
linear velocity and the orientation for the physics engine. The internal
engine tried to find the bounding box for collisions, and used Math.Sign()
on it, which raised the NaN exception.