"UniformGrid" is correctly referenced by "Hammer" when changing scenes.
INFORMATION SECURITY ISSUE
==========================
   As is mentioned multiple times throughout the documentation, the current implementation of passing the "UniformGrid" reference (henceforth referred to as "grid") to the "Hammer" instance (henceforth referred to as "hammer") is not information-secure.
This stems from the fact that the grid is stored in the global injection object "services", which is passed around a plethora of classes.
   Down the line, there might be a code alteration, where some other object manipulates the grid inside the "services" before reaching the hammer, even (in the worst case) going as far as to delete it BEFORE REACHING the hammer.

An irrelevant side note is that a "UniformGrid" instance is newly defined inside the "TreeTutorial.xml" document, something that I had forgotten to do in my previous commits.
Non-existence of a "uniform_grid" XML element (currently) leaves the game to crash.
Ideally, the grid would be constructed automatically by interacting with the physics engine and checking the boundaries of the scene.