A* is run, however route cannot be seen. Read the comment of the commit.
1) The algorithm is not run real time; the game freezes when a call for the hammer to be returned is made.

2) Point 1) makes it difficult to observe what is happenning to the hammer. Visually, it seems like it disappears and then comes from underneath(?) the map. It does stop at the point where it was called from though, which is expected (according to the current implementation) behaviour.

Efficiency consideration
====================
Something to be very careful about is the size of the "uniform_grid" it must be made as small as possible. Even if we can get away with the initialization of the grid --because we can use loading screens-- the same cannot be said about the call to the A* algorithm.
That is because the edges of the graph is RECONSTRUCTED AT EACH CALL according to the mask of the grid, which is considered to be dynamic (see "UniformGrid.FindShortestPathAStar").

Idea for tackling the re-creation of the underlying graph:
Create all the graph edgesonce, and then add/remove only the edges that we care for. This is going to reduce the amount of objects (edges) being (re)created, most of them needlessly so (because they already existed and do not change --e.g. air or static terrain connections).