Effective trick:straight path common -> check for it.Read commit message
PLEA: READ THE WHOLE MESSAGE, UP TO THE END, ESPECIALLY THE LAST COMMENT! IT'S ABOUT OUR GAME DESIGN!

Although this will run in real-time for the majority of player-hammer placement configurations in our current levels, I would not advise to integrate this branch to the main one just yet.
This is because it's going to be a hassle for everyone (+ memory consumption) to write the grids in each scene XML, which takes time (loading screens).
Feel free to integrate it in a new branch, however! :D

Hammer behaviour comments:
==========================
1) If you feel the hammer to be too slow, increase the "hammerSpeed" variable.
2) With the current EventHandlers, the player can catch the hammer MID-AIR (looks cool to me XD)!
3) I'm not quite sure if we wish to have this as a bug or a feature. 
The A* algorithm is an offline algorithm; simply put, this means that once its output is computed, it does not store any of the work done and follows its computed route until the end (unless a new one is computed, which is a whole new A* algorithm execution).
   3a) Algorithms that can handle dynamic movement and uncertain environments do exist (e.g. D* and D*-Lite), but they're not as well documented.
   3b) I don't know if we want to utilize the fact that the player can move around one area while the hammer travels in another (would depend on the speed of the hammer)
   3c) A quick hack so as to forbid scenarios 3b) is to "freeze" the character in place until the hammer gets back to him (maybe standing in a cool "retraction pose" or something).