Move the loop over gameobject.Draw()s into renderer
Logically, the rendering process should be in the hands of the GameRenderer
so it makes sense to move it out of the GameScreen, which is concerned
only with managing the currently loaded scene.

This move also is a step towards implementing shadow maps (where GameRenderer
would need access to the camera and game objects), as well as possibly
more complicated shaders or customised render orders for things like
water (which might need a reflection for game objects). These are all
things that are purely render-specific and should not be in the GameScreen
logic.