Inline the call to SetupDrawTargets and move to within the renderer
It doesn't make sense for GameScreen to be concerned with how the destination
RenderTarget2Ds are set up, so this should be opaquely done with
DrawScene. Since now it's just a small function call within the same
class, it has been inlined.

With this approach, we can change to a different render target
within DrawScene and draw the shadow map too, without GameScreen being
aware of what the state of the intermediate render targets are.