Pass gameTime to all Draw() methods in GameObjects
This will allow shaders and rendering to access the elapsed time and use
it to create varying textures or shading. Primarily, this has been added
for allowing laser shaders to vary over time, since I'm following a
Unity shader tutorial and it has a concept of a Time node to vary UV
coordinates over time.