Implement scrolling laser textures and laser colors
Created a Laser shader effect which essentially copies the main shader
for the calculation on lighting. But it writes nothing to the depth map,
making lasers have no shadows (otherwise it'll have the shadow of the
plane and not the transparent laser part).

The laser pixel shader looks up the material texture (for the colors),
and multiplies it with a scrolling laser texture that defines the alpha
shape. The scroll speed is set in Laser.cs.

Since the current XML setup only allows for one texture per object
(which surely would need to change if we do bump maps or other fancy stuff)
so that slot is for the material texture. The laser texture is assumed
constant and is loaded in the constructor.