Implement querying an alpha mask texture for the laser to soften edges
The new texture is called LaserMask and is softens the horizontal edges
where previously you'd see the billboard effect and it was obvious that
the laser was a 2D plane. This artifact is still somewhat visible if the
laser is scaled to be very wide, but works well for thinner lasers.

Also changed the loading of the laser texture and the laser mask to
correct for the gamma change, since the nice gradients were created in
sRGB. Unless this is corrected, the jump from alpha 0 to alpha 0.1 will
be very harsh compared to 0.9 to 1, and the desired softening effect
won't work.