Re-add the debug entity and grid visualiser, but as a Scene variable
The debug entity and grid visualiser was temporarily removed during the
effort to integrate shaders and implement a rendering system. It has now
been re-added, although it is now under Scene and not the GameScreen.

This move makes sense naturally that the scene controls all GameObjects
and debug objects. It also makes rendering code easy since everything is
already available from the scene variable and we don't have to pass more
arguments.