First attempt at loading screens
- Moves the scene XML initialization into a LoadContent method
- A Scene can now be in a state where it's not loaded, so adds
  appropriate checks everywhere to make sure the scene is loaded
- Scene's OnSceneStart is automatiacally called on the first Update
  after the scene's contents are loaded

The attempt works, and will show the loading screen correctly while
assets get loaded in the background for each level load. Unfortunately,
it is a lot slower (about 10x slower), because texture loads are limited
by the main thread -- one texture load happens per frame, regardless of
how fast the file IO could be, because of OpenGL multi-threading context
restrictions and how ContentManager is implemented. It's faster without
a background thread because doing texture loads inside LoadContent() gets
immediate access to the OpenGL context without waiting for the next frame.

More info:
https://github.com/MonoGame/MonoGame/issues/7991#issuecomment-1500270895
https://community.monogame.net/t/solved-threaded-content-loading-in-mg-3-6/9290/10
https://github.com/MonoGame/MonoGame/blob/1e7b56298ea573765f7f680daec700b1838a8d7b/MonoGame.Framework/Content/ContentManager.cs#L321

The "normal" solution that people apparently do to solve this is to use
their own content managing system that can asynchronously load file data
from a background thread in a batch and queue multiple of them, and then
process them with the OpenGL context in the main thread -- this benefits
from asynchronous file IO and still does Texture2D.SetData in the main
thread, but it's a huge amount of effort for gaining something so small
as 3-4 seconds faster loading...

More info on that:
https://discord.com/channels/355231098122272778/402546044274999299/1103399244372660347