Draw sky first so fog blends with background
This isn't ideal since the sky will now have to be rendered for each
screen pixel even if it'll be occluded by something else, but it's the
only way for the fog to blend into the background near the camera far
plane. Due to the order of rendering, when we render a game object that
needs fog, the background needs to be present already to blend with,
otherwise it'll blend with black.