Fix Pleasing.Tweening.AddProperty not being able to access private fields
Generally it'd be a bad idea (and goes against OOP principles) for
outside classes to access encapsulated private fields, but Tweening is
like a general util function that doesn't really fit into OOP and should
just be able to work with any property we throw at it. We won't be able
to tween private fields at all without this change (even from within the
class itself), so it's a necessary change for Tweening to do anything
useful.