Make GameObject's visual offset property be transformed by local rotations
This change is necessary for the EntityModelOffset (that is tweakable
from the debug UI and the XML) offsets in the correct direction if the
object is rotated. Previously, the offset would be in world coordinates
so the rotation was ignored, resulting in offsets that worked when the
object faced one direction and not in another.