Switch Scene.GameObjects from Dictionary to OrderedDictionary
It turns out that the ordering of a Dictionary is undefined. This is
dangerous for our use case in Scene.GameObjects, since we currently rely
on it being the same as the order defined in the XML, to acheive
consistent save/loads and well-defined render orders.

We switch to an open-source (MIT licensed) OrderedDictionary
implementation, which internally manitains a Dictionary and an index List
to achieve O(1) lookup by key and consistent ordering by insertion order.
This change is transparent except for type changes, since the new
OrderedDictionary implements IDictionary. It supposedly requires slightly
more memory to keep both a dictionary and list in memory, but profiling
revealed unnoticable results so we're fine (our scenes are quite small).

The other options that were considered instead of bringing in a third-
party library are not very good:
- C# *does* come with its own OrderedDictionary, under the
  System.Collections.Specialized namespace, but it's heavily outdated.
  It is implemented using ArrayList (deprecated in favour of List) and
  Hashtable (deprecated in favour of Dictionary), both of which are not
  generic and requires us to store items as `object`s, and box/unbox
  them when we need it. It also does not implement the IDictionary
  interface, lacking in any of the features we rely on like looping over
  KeyValuePair<k,v> using `foreach` or converting to Lists with LINQ.
- SortedDictionary and other Sorted variants sort by the key and not
  insertion order. They are also O(log n) in terms of key access, which
  we want to avoid because we do frequent lookups within scene scripts.
  (Deletion/insertion, not so much)
- Implement one from scratch: too much effort