Implement abiliy to re-order items in the debug scene editor
This wasn't a strict necessity, but makes life much easier for scene
creation and editing. Now that we have shaders that use transparency,
render order has become important. E.g. Lasers (has to be rendered above
most things that could be behind it), Water (has to be rendered after
terrain), or Skyboxes (has to be rendered before terrain).

Currently, the render order is directly mapped to the order of objects
defined in the XML. Unfortunately, because we had no method of
rearranging objects in the level editor, we had to save the XML,
manually change the order of objects in the XML, and load the file again
to test -- this was tedious.

This change implements the ability to just drag and drop to rearrange
items within the level editor. The items at the top render first.