Implement camera frustum dependent shadow maps
An issue with having an absolute definite position for the sunlight like
previously was that you could potentially walk out of the light's view
frustum, where there was no corresponding shadow map entry. This would
result in the player and objects losing shadows outside of a certain
radius from the world-space origin.

This commit implements a moving sunlight based on the camera frustum,
which is how the previous issue is commonly solved. We find the bounds
of the camera's view frustum and use it to create a bounding frustum
from the point of the light, and position our light at the back of that
frustum box, thereby always keeping the camera's view available in the
shadow map.

Unfortunately this causes flickering because subtle changes in the
camera projection introduces pixel-level changes the light frustum, which
leads to pixel level changes in the shadow map, which is sampled at a
resolution low enough for the flickers to be visible on-screen. I'm
still searching for a solution.