Fix frame drop when showing prompts for a non-existent key
The shore scene lagged very much when showing camera movement prompts.
This turned out to be because camera-up and camera-down mappings (which
aren't meaningful to the game but are required to define a
ContinuousUserAction) were assigned to F20 keys that images didn't exist
for. I had been catching KeyNotFoundExceptions and supressed them, but
it turns out that raising exceptions on every frame is very expensive.

The code for handling non-existent keymapping images has been rewritten
using a ContainsKey conditional so we never have exceptions being thrown.