First working attempt at quicker loading screens by extending ContentManager
We make ParseFromXML awaitable, then use the async load method of the
extended content manager to load models and textures. This means that
ParseFromXML runs in an asynchronous thread and is probably not on the
main UI thread. Each LoadAsync also runs in the async threadpool and
may or may not run in their own asynchronous thread.

The LoadAsync method then (in each of those threads) retrieves the file
from the file system (thereby achieveing asynchronous IO), and then uses
ScriptUtils to wait for the next update, and continue the loadnig code
synchronously within an Update() in the main thread.

It's largely undocumented and debug logs scattered throughout, I'll get
back to it in a few commits time.