Switch menu from using Myra's Menu widget to use just a VerticalStackPanel
We had been using Myra's Menu widget for the title and pause screens.
This widget can contain an arbitrary number of MenuItems and
MenuSeparators. While these were fine for the title and pause screens,
we now need to add an options screen too, and that will contain more
complex functionality than just clickable labels. Ideally, we want to
have arbitrary elements as MenuItems.

MenuItems implement the IMenuItem interface, so I thought of creating
new classes that would implement IMenuItem and would be extensive in
functionality. Unfortunately, there are a lot of hardcoded sections in
Myra's source code that do something like:

IMenuItem something;
if (something is MenuItem) { // treat it as just a label }
else { // assume it is a separator and cast it to one }

The solution I've implemented here is to replicate the entire
functionality of Menu and MenuItems with just a VerticalStackPanel and
click/hover events. So far the change looks transparent on the UI-side
and the keyboard/controller/mouse interactions are identical. But it now
accepts a List<Widget> for the menu instead of List<MenuItem>, which
means we can display any sort of menu! I think sliders for volume and
some toggles for fullscreen are core necessities.