Make sure every change to HoverIndex in UI also changes Desktop.FocusedKeyboardWidget
HoverIndex is used for showing which item is currently hovered/selected.
But Myra sends keyboard events (like OnKeyDown) only to the internally
focused widget, which must stay in sync with HoverIndex. This is because
we will be implementing some widgets which take left/right key inputs to
toggle between values, and those widgets need to receive keyboard events.

This commit makes sure that every change in HoverIndex also modifies
Myra's focus.