Implement an options screen for controlling volume and full screen
The options don't persist yet nor are they fully utilised yet (e.g.
music volume gets overwritten in GameScreen so it's useless right now).

This commit implements an options screen, a screen derived from the
AbstractMenuScreen. It has a smaller font size to accommodate the
probably growing list of options. A toggle-style option widget is
implemented using HorizontalStackPanel and two labels. It supports ints
as the internal value, but a wrapper function has been created to easily
create various options as long as we give it a function to convert it to
a string, a minimum/maximum value, and a function to apply the change.