fix: screen manager no longer gievs input focus to screens that are mid-transition
This fixes an issue with how you could quickly click twice to trigger on
some menu item to trigger its handler twice, because the new screen was
still mid-transition and passed focus through. The fix is to make the
screen above steal focus even during transitions (basically as long as
some part of it is drawn), but not give the hasFocus boolean to it unless
it has finished transitioning.