Fix Pleasing (Tweening) library bug that made all setter function tweens not work
Apparently this library we're using isn't perfect. It was not adding any
of the tweens defined using setter functions (instead of property names)
to the update list, so they didn't get updated.

The original library hasn't been updated for several years and the author
is unresponsive so this fix is going in the code directly here.