Prevent duplicate screens from being added to the screen stack
This change makes sure no two identical screen instances can be added
to the screen stack. This fixes bugs where you could double click a
menu item really quickly and it would create two new screens on the
stack, and one of them would properly get removed by RemoveScreen
while the other wouldn't since RemoveScreen only removes the first
instance.

The other alternative would be to allow multiple copies of the same
screen instances in the stack, but then we need a good method to
prevent quick double-clicks from adding two screens, and that is a
lot of hassle.