Move SetFullScreen logic into SetResolution since it's a special case of resolution change
Because we're using software full screens, toggling full screen has a
two-step process of signaling to the OS plus executing resolution-change
code. This means SetFullScreen is a special case of SetResoltion with
some fluff to signal to the OS.

It was originally implemented as a wrapper function, but there arose a
problem: when exiting full screen back to a window, the window size was
not known to the SetFullScreen function, and it couldn't call
SetResolution with the appropriate arguments. This change merges the two
functions into one so that full screen logic has access to the window
width and height when necessary.