Use the result of GPU.Viewport for rendering after setting back buffer
Macs in particular have a tendency to not actually set the requested
back buffer size, so in order to render to the actual resolution we
should use gpu.Viewport instead of the requested back buffer sizes.
DPI scaling can also mess with this, apparently, although this wasn't
the reason for this commit.

More info: https://github.com/MonoGame/MonoGame/issues/4802#issuecomment-577090157