Make a faster, accelerating hammer. See message for eminent bug.
This has a bug with the rock boulders. The rock boulders acquire the velocity of the hammer at the first frame of contact. While this was fine when the hammer had a constant speed, now the rock boulder and the hammer reach the player heterochronously, unless --of course-- the hammer has already reached its maximum, constant velocity.
Contact Sid regarding this.