Rock boulders acquire the velocity of the hammer. Sound replaying issue.
With this slight change in code, now the rocks have the same velocity (not just speed!) with the hammer at all frames. One issue (not destructive, just annoying to the ear) is that I haven't been able to figure out how to make the sound of the rock being hit only once. That being the case, we could theoretically mask it as "this is the sound effect of the rock being dragged" or something.
Requires a bit more testing to see whether it creates issues.