Fix WaterBoundsObject collisions after tree dust particles interact with it
The particles added by the particle systems in recent commits are in the
physics space and can interact with objects. I'm not exactly sure why,
but WaterBoundsObject don't like it, sometimes going back to Defer
collision handling (when it should be NoSolver) after a particle ends its
collision with the bound, despite there being a tree or rock collision
ongoing.

This commit makes sure that we explicitly set solver state to NoSolver if
there are more than one tree/rock collision, so we can now handle the
scenario where the solver state goes to Defer for whatever reason when
dealing with particles.