Fix particles not receiving shadow map parameters from parent
Particles used to receive those parameters unintentionally because the
same effect instance was shared between the parent object and its
particles, and the parameter value was never modified between the two
Draw calls. But this was sort of undefined behaviour. Since we now
clone effects to have a unique instance for each object, this shadow map
parameter needs to be passed in explicitly manually.