Remove janky player jump when going over rocks in water
The new stepping mechanism for the player implemented in commit 940c0b8
makes the old janky player teleportion for rocks unnecessary.
Essentially, players will automatically climb up any small difference in
height between the ground and the floating rock, which means we won't
fall into the water, and we don't have to manually teleport the user up
high (which was visually janky).

We keep the logic for the rock-on-rock situation (and rock-on-tree),
since the moving rocks don't have a stepping mechanism yet.