Fine-tuned uniform grid for "PrototypePuzzle". Read commit comment.
The increase in loading time is noticeable (requires 3-4 seconds to load on my laptop). This is because the resolution of the grid has been increased.

Why do this? I remember a lot of players struggling with this level during the public presentation of the game, because the hammer would try to follow the Bezier Curves, in which mode (currently) the hammer does not drag rocks.

Further improvements?
The loading time and memory overhead *can* be reduced however. This can be done by completely disallowing the player to go to the useless areas of the map; those are the empty areas past the "forests" at the beginning of the level.  Currently, as I cannot exclude the possibility of the player reaching those points, I have to account for those positions. However, those positions are not only unnecessary for the level, but are also the edges of the scene. Therefore, if access to those empty areas becomes impossible, then the grid may be reduced.